solution/services/media-service/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
cachetools
httpx
python-multipart
